+++ src/structures/tag.rs
@@ -7,7 +7,7 @@ pub struct Category {
     /// The name of the category
     pub name: String,
     /// The project type this category is applicable to
-    pub project_type: project_structs::ProjectType,
+    pub project_type: project::ProjectType,
 }
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
@@ -17,7 +17,7 @@ pub struct Loader {
     /// The name of the loader
     pub name: String,
     /// The project types that this loader is applicable to
-    pub supported_project_types: Vec<project_structs::ProjectType>,
+    pub supported_project_types: Vec<project::ProjectType>,
 }
 
 #[derive(Deserialize, Serialize, Debug, Clone)]